requirements.txt
Pillow==10.4.0
pynput==1.7.7
google-api-python-client==2.139.0
numpy==2.0.1
pyautogui==0.9.54
python-dotenv==1.0.1
requests==2.32.3